refactor: upgrade dev page
